add Game, GameBoard and Player files
Create basic functionallity for the Game where you have to type in
yourself which column you want to populate.

Player can make a move which invokes GameBoard.update_board which
checks if the movement is valid or not and updates the correctly
column index.

Not Implemented yet:
Machine(Player), is_game_over, interactive game and other stuff.